requirements.txt
fastapi==0.111.0
Flake8-pyproject==1.2.3
taskipy==1.12.2
SQLAlchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic-settings==2.2.1
fastapi-pagination==0.12.24